backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg[binary]>=3.2,<3.3
chromadb==0.4.18
openai==1.3.5
google-generativeai==0.3.1
pypdf==4.0.1
python-dotenv==1.0.0
requests==2.31.0